started porting over to typescript...
