+++ .babelrc
@@ -1,0 +1,4 @@
+{
+    "presets": ["env"],
+    "plugins": ["transform-object-rest-spread"]
+}
