+++ index.ts
@@ -1,0 +1,1 @@
+export {StaticSiteGenerator} from './src/StaticSiteGenerator';
