+++ package.json
@@ -1,0 +1,35 @@
+{
+  "name": "jss-static-site-generator",
+  "version": "1.0.0",
+  "description": "A static site generator for Sitecore JSS applications. Read more and find information on getting started [here](https://www.adamlamarre.com/static-site-generator-for-sitecore-jss/).",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "clean": "del-cli dist types",
+    "build": "npm run clean && tsc"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/erzr/jss-static-site-generator.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/erzr/jss-static-site-generator/issues"
+  },
+  "homepage": "https://github.com/erzr/jss-static-site-generator#readme",
+  "devDependencies": {
+    "@sitecore-jss/sitecore-jss-dev-tools": "^12.0.0",
+    "@types/graphql": "^14.2.3",
+    "@types/node": "^12.7.1",
+    "del-cli": "^1.1.0",
+    "typescript": "^3.1.3"
+  },
+  "dependencies": {
+    "apollo-cache-inmemory": "^1.3.12",
+    "apollo-client": "^2.4.8",
+    "apollo-link-batch-http": "^1.2.6",
+    "graphql": "^14.0.2",
+    "graphql-tag": "^2.10.0"
+  }
+}
