+++ src/LayoutService.ts
@@ -1,0 +1,3 @@
+export default interface LayoutService {
+    
+}
