+++ tsconfig.json
@@ -1,0 +1,36 @@
+{
+    "compilerOptions": {
+      "target": "es5",
+      "module": "commonjs",
+      "newLine": "LF",
+      "emitDecoratorMetadata": true,
+      "experimentalDecorators": true,
+      "forceConsistentCasingInFileNames": true,
+      "lib": [
+        "es2017",
+        "dom"
+      ],
+      "moduleResolution": "node",
+      "noFallthroughCasesInSwitch": true,
+      "noImplicitReturns": true,
+      "noUnusedLocals": true,
+      "noUnusedParameters": true,
+      "strict": true,
+      "typeRoots": [
+        "node_modules/@types"
+      ],
+      "outDir": "dist",
+      "declaration": true,
+      "declarationDir": "./types",
+      "allowSyntheticDefaultImports": true,
+      "esModuleInterop": true
+    },
+    "exclude": [
+      "node_modules",
+      "types",
+      "dist",
+      "src/customTypes",
+      "src/**/*.test.ts",
+      "src/testData/*.ts"
+    ]
+  }
